Dental-Teeth/DentalScanner/DentalScanner/requirements.txt
flask==2.3.3
requests==2.32.5
inference_sdk==0.56.0
python-dotenv>=1.0
